requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
backports.asyncio.runner==1.2.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.1.8
dnspython==2.7.0
email-validator==2.3.0
exceptiongroup==1.3.1
fastapi==0.128.2
fastapi-cli==0.0.20
fastapi-cloud-cli==0.11.0
fastar==0.8.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
iniconfig==2.1.0
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
nba_api==1.10.2
numpy==2.0.2
packaging==26.0
pandas==2.3.3
pluggy==1.6.0
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic-extra-types==2.11.0
pydantic-settings==2.11.0
pydantic_core==2.41.5
Pygments==2.19.2
pytest==8.4.2
pytest-asyncio==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
rich==14.3.2
rich-toolkit==0.18.1
rignore==0.7.6
sentry-sdk==2.52.0
shellingham==1.5.4
six==1.17.0
starlette==0.49.3
tomli==2.4.0
typer==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.39.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
